+++ level_sensor/simulation.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <stdint.h>
 
-uint16_t readSimulatedRaw();
+int32_t readSimulatedRaw();
 void setSimulationMode(uint8_t mode);
